+++ index.js
@@ -17,7 +17,18 @@ program
   .requiredOption('-p, --path <path>', 'path/url to swagger scheme')
   .option('-o, --output <output>', 'output path of typescript api file', './')
   .option('-n, --name <name>', 'name of output typescript api file', 'Api.ts')
-  .option('-d, --default-as-success', 'use "default" response status code as success response too.\n' + 'some swagger schemas use "default" response status code as success response type by default.', false)
+  .option(
+    '-d, --default-as-success',
+    'use "default" response status code as success response too.\n' +
+    'some swagger schemas use "default" response status code as success response type by default.',
+    false
+  )
+  .option(
+    '-r, --responses',
+    'generate additional information about request responses\n' +
+    'also add typings for bad responses',
+    false,
+  )
   .option('--route-types', 'generate type definitions for API routes', false)
   .option('--no-client', 'do not generate an API class', false);
  
@@ -30,6 +41,7 @@ const {
   routeTypes,
   client,
   defaultAsSuccess,
+  responses,
 } = program;
 
 generateApi({
@@ -38,6 +50,7 @@ generateApi({
   generateRouteTypes: routeTypes,
   generateClient: client,
   defaultResponseAsSuccess: defaultAsSuccess,
+  generateResponses: responses,
   input: resolve(process.cwd(), path),
   output: resolve(process.cwd(), output || '.')
 })