+++ src/config.js
@@ -1,6 +1,8 @@
 
 const config = {
   /** CLI flag */
+  generateResponses: false,
+  /** CLI flag */
   defaultResponseAsSuccess: false,
   /** CLI flag */
   generateRouteTypes: false,