+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 Breaking Changes:  
   - Rename default typescript api file name (prev `api.ts`, now `Api.ts`)  
 Features:  
-  - Improve response body type definitions  
-  - Response declarations in request rescription  
-  - Types for bad responses  
   - `-d, --default-as-success` option. Allows to use "default" status codes as success response type  
+  - `-r, --responses` option. Response declarations in request rescription  
+    This option adds comments of the possible responses from request  
+    ![responses comments](./assets/changelog_assets/responses-comments.jpg)  
+    Also typings for `.catch()` callback  
+    ![responses catch types](./assets/changelog_assets/responses-catch-types.jpg)  
+  - Improve response body type definitions  
+  - Types for bad responses  
 
 # 1.3.0  
 Features: